arcpro/deploy/requirements.txt
# ArcGIS Pro到QGIS迁移系统依赖包
# 核心依赖
psycopg2-binary>=2.9.5
pyyaml>=6.0
lxml>=4.9.0
psutil>=5.9.0

# 开发依赖（可选）
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
bandit>=1.7.5

# GIS相关（根据环境选择）
# arcpy - 仅在ArcGIS Pro环境可用，需单独安装
# qgis - 仅在QGIS环境可用，需单独安装
# gdal - 通过系统包管理器安装

# 注意：arcpy和qgis通常通过GIS软件安装包提供，不通过pip安装